morguard/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
cryptography==42.0.5
psycopg2-binary==2.9.7
pdfplumber==0.11.0
adrf==0.1.6
gunicorn==22.0.0
uvicorn==0.29.0
llama-index==0.10.40
boto3
python-dotenv==1.0.1
drf-yasg==1.21.7
PyJWT==2.8.0
sqlalchemy[asyncio] 
pgvector
asyncpg